+++ app/src/main/java/com/infinity/omos/data/UserToken.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 data class UserToken(
     @SerializedName("accessToken") val accessToken: String?,
     @SerializedName("refreshToken") val refreshToken: String?,
-    @SerializedName("userId") val userId: Long
+    @SerializedName("userId") val userId: Int
 )
